Remove tests module from lib.rs to streamline codebase and improve organization. The tests have been relocated for better structure and maintainability.
